Move CLI config processing out from under registry dir
No logic changes should be in here, just moving things around.
